Introduce Separate Builder (+ Moved Material files)
